[feat] add find Board Cache by discordId & refactoring
캐시에서 디스코드 아이디를 통해 board를 검색하는 메서드 추가
이외 리팩토링
