+++ src/main/java/space/space_spring/domain/post/application/port/in/boardCache/AddBoardCacheUseCase.java
@@ -1,4 +1,4 @@
-package space.space_spring.domain.post.application.port.in;
+package space.space_spring.domain.post.application.port.in.boardCache;
 
 public interface AddBoardCacheUseCase {
     boolean add(Long boardId);
