+++ src/main/java/space/space_spring/domain/post/application/service/CreateBoardService.java
@@ -3,7 +3,7 @@ package space.space_spring.domain.post.application.service;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import space.space_spring.domain.post.application.port.in.AddBoardCacheUseCase;
+import space.space_spring.domain.post.application.port.in.boardCache.AddBoardCacheUseCase;
 import space.space_spring.domain.post.application.port.in.createBoard.CreateBoardCommand;
 import space.space_spring.domain.post.application.port.in.createBoard.CreateBoardUseCase;
 import space.space_spring.domain.post.application.port.out.CreateBoardPort;
@@ -29,7 +29,7 @@ public class CreateBoardService implements CreateBoardUseCase {
         if(addBoardCacheUseCase.add(boardId)){
             throw new CustomException(BOARD_NOT_FOUND, BOARD_NOT_FOUND.getMessage()+": redis 저장과정에서 찾을 수 없습니다");
         }
-        
+
         return boardId;
     }
 }
